Syntax now requires 'clone' subcommand; accepts a list of branch specs to fetch
For future expansion, a 'clone' subcommand must now be given. The repo dir is derived
from the Perforce path if not specified. A set of "branch specs" are accepted to map
subdirectories of the Perforce path to branches in refs/remotes/p4 in GIT. The base
path and branch specs are written to the GIT repository's configuration for future
use.

Passing "-d" as a parameter before the subcommand activates $PerforceLink::DEBUG.