Don't try to suck all changes down at once.
Although that would make things a lot faster, it doesn't work for initial
imports for us: there's just too much under //depot/site/... for it to scan
without restricting it to main or a specific release.

So the code now in does "changes" commands for each branch root, and then
combines the list itself. Later I'll add a "fast" mode that does it the quick
way, for incremental imports.