Move most logic to a PerforceLink::Git package.
This helps ease the crazy passing around of parameters that was getting more
and more painful. Also things like arranging the pipe to git fast-import are
now inside the fetch_p4_changes call rather than being duplicated in the
script.

When piping to fast-import, we maintain a marks file inside the target
repository.

It is now possible to use the 'fetch' command with an output file (and the
configuration's idea of the current changelist is not updated in this case).