Deal with branches yet to be used differently when doing a slow change scan.
If you do something like "p4 changes
//depot/project/branch/foo/...@10000,#head" and the branch does not in fact
have any changes for 100,000 or so changelists since then, you may get a
maxscanrows error.

Doing "p4 changes //depot/project/branch/foo/..." always seems to be OK, and
also "p4 changes //depot/project/branch/foo/...@10000". So use a combination
of the latter to determine if there are any old changes (that we are
skipping), and if there are not, then use the former to just get all changes
on the branch.